refactor: 컴포넌트 조작 메서드를 분리
