+++ src/components/AddRestaurant/handleAddRestaurant.ts
@@ -1,0 +1,9 @@
+import IRestaurant from "../../type/IRestaurant";
+import RestaurantList from "../RestaurantList";
+
+export const addNewRestaurant = (newRestaurant: IRestaurant) => {
+  const restaurantList = document.getElementById("restaurantList");
+  if (restaurantList instanceof RestaurantList) {
+    restaurantList.addRestaurant(newRestaurant);
+  }
+};
